src/requirements.txt
aiofiles==22.1.0
alembic==1.11.3
antlr4-python3-runtime==4.9.3
anyconfig==0.10.1
anyio==3.5.0
appnope==0.1.2
arrow==1.2.3
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
binaryornot==0.4.4
black==23.9.1
blinker==1.6.2
Bottleneck==1.3.5
build==0.10.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colourmap==1.1.16
comm==0.1.4
contourpy==1.0.5
cookiecutter==2.3.0
cryptography==41.0.2
cycler==0.11.0
databricks-cli==0.17.7
datazets==0.1.8
deap==1.4.1
debugpy==1.8.0
decorator==5.1.1
docker==6.1.3
dynaconf==3.2.0
entrypoints==0.4
executing==0.8.3
fastapi==0.95.2
Flask==2.3.3
fonttools==4.25.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
graphql-core==3.2.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.19.2
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kedro==0.18.12
kedro-mlflow==0.11.9
kedro-viz==6.5.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.0
mlflow==2.7.1
more-itertools==9.0.0
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numexpr==2.8.4
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.9.5
packaging==23.0
pandas==1.5.3
parse==1.19.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pip-tools==7.3.0
platformdirs==3.11.0
plotly==5.16.1
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2022.7
PyYAML==6.0
pyzmq==25.1.1
querystring-parser==1.2.4
requests==2.31.0
rich==13.5.1
rope==1.9.0
scatterd==1.3.7
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.11.2
secure==0.3.0
semver==3.0.1
setuptools==67.8.0
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.2.0
starlette==0.27.0
stopit==1.1.2
strawberry-graphql==0.205.0
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toposort==1.10
tornado==6.2
TPOT==0.12.1
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.7.1
update-checker==0.18.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.20.0
watchgod==0.8.2
wcwidth==0.2.5
websocket-client==0.58.0
websockets==11.0.3
Werkzeug==2.3.7
wheel==0.38.4
xgboost==1.7.3
zipp==3.11.0
